+++ lib/button.rb
@@ -1,3 +1,4 @@
 require 'button/client'
+require 'button/constants'
 require 'button/version'
 require 'button/utils'