+++ lib/button/constants.rb
@@ -1,0 +1,6 @@
+module Button
+  module Constants
+    TIME_FIELD_CREATED = 'created_date'.freeze
+    TIME_FIELD_MODIFIED = 'modified_date'.freeze
+  end
+end