+++ lib/button/resources/transactions.rb
@@ -3,23 +3,9 @@ require 'button/resources/resource'
 module Button
   # https://www.usebutton.com/developers/api-reference/
   #
-  class Accounts < Resource
-    def path(account_id = nil)
-      return "/v1/affiliation/accounts/#{account_id}/transactions" if account_id
-      '/v1/affiliation/accounts'
-    end
-
-    # Gets a list of available accounts
+  class Transactions < Resource
+    # Gets a list of transactions
     #
-    # @return [Button::Response] the API response
-    #
-    def all
-      api_get(path)
-    end
-
-    # Gets a list of transactions for an account
-    #
-    # @param [String] account_id the account id to look up transactions for
     # @option [String] cursor the account id to look up transactions for
     # @option [ISO-8601 datetime String] start The start date to filter
     #   transactions
@@ -28,14 +14,14 @@ module Button
     # @option [String] time_field time field start and end filter on
     # @return [Button::Response] the API response
     #
-    def transactions(account_id, opts = {})
+    def all(opts = {})
       query = {}
       query['cursor'] = opts[:cursor] if opts[:cursor]
       query['start'] = opts[:start] if opts[:start]
       query['end'] = opts[:end] if opts[:end]
       query['time_field'] = opts[:time_field] if opts[:time_field]
 
-      api_get(path(account_id), query)
+      api_get('/v1/affiliation/transactions', query)
     end
   end
 end